Create a CompileOnly spec for props edge cases.
